initiated a new django project

+++ todoapp/tasks/views.py
@@ -1,7 +1,5 @@
-from django.shortcuts import render
-from django.http import HttpResponse, HttpResponsePermanentRedirect
-from django.urls import reverse
-
+from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponse, HttpResponseNotAllowed
 from .models import Todotask
 from .forms import TaskForm
 
@@ -23,7 +21,6 @@ def process_task(request):
         if task.is_valid():
             task.save()
     context = {'task':task}
-    # return HttpResponse("Task has been added successfully")
     return render(request,'add_task.html',context)
 
 
@@ -35,7 +32,20 @@ def task_details(request,pk=None):
     task = {'task':task}
     return render(request,'task_details.html',task)
 
-def delete(request,pk):
-    task = Todotask.objects.get(pk=pk)
-    task.delete()
-    return HttpResponse('Deleted')
+def delete(request,pk=None):
+    if request.method == 'POST':
+        if pk:    
+            task = get_object_or_404(Todotask,id=pk)
+            task.delete()
+        return redirect('home')
+    else:
+        return HttpResponseNotAllowed(['POST'])
+
+def edit(request,pk=None):
+    if pk:
+        task = Todotask.objects.get(pk=pk)
+    else:
+        task = ''
+    task = {'task':task}
+    return render(request,'task_edit.html',task)
+
